Initialize Limiting Factor crate
Summary:
The limiting-factor crate offers facilities to implement a REST API.

The goal of this library is to provide:
  - glue code for Rocket and Diesel
  - standard API replies
  - boilerplate to parse environment and run a server

That allows an API or a back-end web server to focus on requests
and data model, and to maintain helper methods as a separate library.

*Etymology*

Limiting Factor is a (D)GOU in The Player of Games, a Culture novel.

Test Plan: `cargo build`

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1658