Implement ApiJsonResponse for Serialize and Json
Summary:
Generate a JSON output from an object ready to be serialized to JSON
is trivial. But to be coherent with the more complex objects, we can
offer `into_json_response()` through a ApiJsonResponse implementation.

The same applies for the Json itself.

Test Plan: ApiJsonResponse for a Serialize is used by D1868.

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1869