Update for Rocket 0.4 and recent Rust
Summary:
Rocket 0.4 allows to use directly a Status as a response,
without any need to wrap in a Failure construct. This simplifies
the code and is more intuitive. As such, we've some wrapper methods
to deprecate.

Compiler plugins for code generation is now stabilized,
it's not feature-gated anymore.

Test Plan:
- compile the lib with Rust 1.33.0-nightly (a2b0f247b 2018-12-30)
- compile a project against the library

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2048