Update build instructions for PostgreSQL on Windows
Summary:
The pq-sys offers rust bindings for libpq. It provides a build script
to detect the path of the PostgreSQL library.

To directly provide the relevant directory, LIB isn't used anymore,
but a specific variable is, PQ_LIB_DIR.

If this variable changes, the build script is run again. As such,
`cargo clean` isn't needed, which avoid to recompile dozen of crates.

Test Plan: Compiled with nightly-x86_64-pc-windows-gnu toolchain.

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2425