Prepare the repository for current state of art
Summary:
Per T1231, a dichotomic search highlighted the last nightly toolchain able to
build Rocket v0.4 is nightly-2024-12-15. This change adopts that version to
generate documentation through Jenkins pipeline and documents it in README.

As we want to support more recent frameworks, a modular structure is required:
  - different crates by framework can be provided
  - common patterns can go to a core crate

To adopt this repository structure, the current code is moved to rocket-legacy/

In next commits, framework-agnostic code will be moved to the core crate.

Ref T2131 - required toolchain version
Ref T2135 - repository structure

Test Plan: `cargo build`

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T2135, T2131

Differential Revision: https://devcentral.nasqueron.org/D3713